+++ cogment/utils.py
@@ -21,15 +21,28 @@ from cogment.version import __version__
 import logging
 import types
 
+logger = logging.getLogger("cogment.sdk")
+
 
 def _logger_trace(self, msg, *args, **kwargs):
     """Simplified trace log output method"""
     self.log(self.TRACE, msg, *args, **kwargs)
 
 
-logger = logging.getLogger("cogment.sdk")
 setattr(logger, "TRACE", 5)
 setattr(logger, "trace", types.MethodType(_logger_trace, logger))
+
+
+def _logger_deprecated(self, msg):
+    """Log deprecation as warning but only once"""
+    if msg not in self.DEPRECATED_LOGGED:
+        self.warning(msg)
+        self.DEPRECATED_LOGGED.add(msg)
+
+
+setattr(logger, "DEPRECATED_LOGGED", set())
+setattr(logger, "deprecated", types.MethodType(_logger_deprecated, logger))
+
 logger.setLevel(logging.INFO)