+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Reward tick ID was hard coded to -1
 
+### Changed
+
+- Deprecation warning are only logged once
+
 ### Added
 
 - Support for actor diconnection and response timeout parameters